Utility function to run validation before submit review
